refactor(script): extract parser function in its own file for testing
